Refactored IncludeAdditionalQuery into new utils.py file
